Allow trainArrow's plots to be generated offline
Currently the summary plots generated by Arrow, specifically
for the Transition probabilities, require data not stored in
the "cleaned" fit.  This is means that any basic error caused
by new a model-form, or investigations into the results, can't
be done without wastefully re-running the whole training.

This fix saves the relevant data from the "raw" fit and directs
the plotting functions to use it (from the "cleaned" fit) instead.
The result is a slightly larger "fit.rda", but one for whom plots
are reproducible and interrogatable.